ci: bump GOVERSION and fix golint
GitHub Actions was configured to use Go 1.17, which is quite old. This commit
bumps it to Go 1.21, which is the older of the two currently-supported Go
versions.

This commit also adjusts the command-line for installing `golint` to be
compatible with more recent Go versions.

Keeping go.mod at 1.17 for now though, unless/until we actually need newer Go
language features. Since github.com/skeema/knownhosts is imported by a lot of
other Go packages, we don't want to force increases to that listed version
unnecessarily.