comments: update to reflect golang.org/x/crypto/ssh/knownhosts change
As of the next release of the golang.org/x/crypto module, it will be possible
for golang.org/x/crypto/ssh/knownhosts to return multiple keys of the same
type for the same host, per https://github.com/golang/crypto/pull/254. We
already handle that gracefully since v1.2.1 (see commit 3a35d9f), but the code
comment describing this situation needed to be updated.